backend/requirements.txt
# ── Web Framework ─────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# ── Database ──────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pgvector==0.3.5

# ── Auth ──────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ── AWS ───────────────────────────────────────────
boto3==1.35.80
botocore==1.35.80

# ── Redis ─────────────────────────────────────────
redis[asyncio]==5.2.1

# ── Pydantic / Config ─────────────────────────────
pydantic==2.10.3
pydantic-settings==2.6.1

# ── Template Rendering ────────────────────────────
jinja2==3.1.4

# ── Token Counting ────────────────────────────────
tiktoken==0.8.0

# ── PDF Processing ────────────────────────────────
pypdf2==3.0.1
weasyprint==63.0

# ── Utilities ─────────────────────────────────────
python-dotenv==1.0.1
httpx==0.28.1
tenacity==9.0.0
structlog==24.4.0
numpy==1.26.4

# ── Dev / Test ────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1
ruff==0.8.3
